Вносит правки в модулях
