+++ js/data.js
@@ -1,5 +1,3 @@
-import {getRandomNumber, getUniqRandomNumber} from './utils.js';
-
 const COMMENT_MESSAGES = [
   'Всё отлично!',
   'В целом всё неплохо. Но не всё.',
@@ -19,24 +17,4 @@ const COMMENT_USERS = [
 
 const PHOTO_DESCRIPTIONS_COUNT = 25;
 
-const uniqCommentId = getUniqRandomNumber(20);
-
-const generateComment = () => ({
-  id: uniqCommentId(),
-  avatar: `img/avatar-${getRandomNumber(1, 6)}.svg`,
-  message: COMMENT_MESSAGES[getRandomNumber(0, COMMENT_MESSAGES.length - 1)],
-  name: COMMENT_USERS[getRandomNumber(0, COMMENT_USERS.length - 1)],
-});
-
-const generatePhotoDescription = (id) => ({
-  id,
-  url: `photos/${id}.jpg`,
-  description: 'Фотография пряников на деревянном столе. Но это не точно.',
-  likes: getRandomNumber(15, 200),
-  comments: Array.from({length: getRandomNumber(1, 2)}, generateComment),
-});
-
-const photoDescriptions = Array.from(
-  {length: PHOTO_DESCRIPTIONS_COUNT},
-  (_, index) => generatePhotoDescription(++index)
-);
+export {COMMENT_MESSAGES, COMMENT_USERS, PHOTO_DESCRIPTIONS_COUNT};
